backend/requirements.txt
fastapi
uvicorn
requests
unidecode
python-dotenv